+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   * all `stubbles\peer\http\HttpResonse` methods now throw a `stubbles\peer\ProtocolViolation` when the response can not be read properly
   * `stubbles\peer\Timeout` now extends `stubbles\peer\ConnectionFailure`
   * `stubbles\peer\Uri::hasDefaultPort()` will now return `true` instead of `false` when the string it was constructed from doesn't contain a port
+  * deprecated `stubbles\peer\http\AcceptHeader::getSharedAcceptables()`, use `stubbles\peer\http\AcceptHeader::sharedAcceptables()` instead, will be removed with 9.0.0
 
 
 7.1.0 (2016-06-08)